fix: rename logging module to avoid shadowing standard library
