+++ gpustack/cmd/prerun.py
@@ -8,7 +8,7 @@ from typing import List, Dict
 from gpustack.config.config import Config
 from gpustack.schemas.config import GatewayModeEnum
 from gpustack.envs import MIGRATION_DATA_DIR, DATA_MIGRATION
-from gpustack.logging import setup_logging
+from gpustack.log import setup_logging
 from gpustack.cmd.start import (
     start_cmd_options,
     parse_args,
