+++ gpustack/exporter/exporter.py
@@ -9,7 +9,7 @@ from prometheus_client.core import (
 import uvicorn
 from gpustack.config.config import Config
 from gpustack.exporter.bus_metrics import BusMetricsCollector
-from gpustack.logging import setup_logging
+from gpustack.log import setup_logging
 from gpustack.schemas.config import ModelInstanceProxyModeEnum
 from gpustack.schemas.clusters import Cluster
 from gpustack.schemas.models import Model
