+++ gpustack/worker/exporter.py
@@ -12,7 +12,7 @@ from prometheus_client.core import (
 )
 from gpustack.client.generated_clientset import ClientSet
 from gpustack.config.config import Config
-from gpustack.logging import setup_logging
+from gpustack.log import setup_logging
 from gpustack.utils.name import metric_name
 from gpustack.worker.collector import WorkerStatusCollector
 import uvicorn
